+++ packages/datepicker/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.0.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datepicker@9.0.4...@pluralsight/ps-design-system-datepicker@9.0.5) (2020-11-30)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-datepicker
+
+
+
+
+
+## [9.0.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datepicker@9.0.3...@pluralsight/ps-design-system-datepicker@9.0.4) (2020-11-24)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-datepicker
+
+
+
+
+
 ## [9.0.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datepicker@9.0.2...@pluralsight/ps-design-system-datepicker@9.0.3) (2020-11-18)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-datepicker