+++ packages/linearprogress/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.14](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-linearprogress@5.2.13...@pluralsight/ps-design-system-linearprogress@5.2.14) (2020-11-30)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-linearprogress
+
+
+
+
+
+## [5.2.13](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-linearprogress@5.2.12...@pluralsight/ps-design-system-linearprogress@5.2.13) (2020-11-24)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-linearprogress
+
+
+
+
+
 ## [5.2.12](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-linearprogress@5.2.11...@pluralsight/ps-design-system-linearprogress@5.2.12) (2020-11-18)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-linearprogress