+++ packages/row/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.0.1](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-row@11.0.0...@pluralsight/ps-design-system-row@11.0.1) (2020-11-30)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-row
+
+
+
+
+
+# [11.0.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-row@10.1.14...@pluralsight/ps-design-system-row@11.0.0) (2020-11-24)
+
+
+* Feat/drawer rc conversion (#1472) ([01cc3ce](https://github.com/pluralsight/design-system/commit/01cc3ce200afa8117eca187c56a658425bd5f1be)), closes [#1472](https://github.com/pluralsight/design-system/issues/1472)
+
+
+### BREAKING CHANGES
+
+* see docs
+
+
+
+
+
 ## [10.1.14](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-row@10.1.13...@pluralsight/ps-design-system-row@10.1.14) (2020-11-18)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-row