+++ packages/screenreaderonly/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.14](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-screenreaderonly@2.1.13...@pluralsight/ps-design-system-screenreaderonly@2.1.14) (2020-11-30)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-screenreaderonly
+
+
+
+
+
+## [2.1.13](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-screenreaderonly@2.1.12...@pluralsight/ps-design-system-screenreaderonly@2.1.13) (2020-11-24)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-screenreaderonly
+
+
+
+
+
 ## [2.1.12](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-screenreaderonly@2.1.11...@pluralsight/ps-design-system-screenreaderonly@2.1.12) (2020-11-18)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-screenreaderonly